create ZS state ZS_FirePlaceLighter
Npc in this ZS state will detect fireplaces. If fireplace is not lit, if Npc has torch in hand / in inventory and it is close enough, it will go and lit it. Custom perception B_FirePlaceLighter_AssessSelf will be checking for nearby fireplaces even if Npc AI is 'busy' walking.